@@ -11,7 +11,7 @@ public interface UserRepository extends JpaRepository<User,Integer> {
     Optional<User> findByUsername(String username); 
     
     @Modifying()
-    @Query("update User u set u.firstname=:firstname, u.lastname=:lastname, u.country=:country where u.id = :id")
-    void updateUser(@Param(value = "id") Integer id,   @Param(value = "firstname") String firstname, @Param(value = "lastname") String lastname , @Param(value = "country") String country);
+    @Query("update User u set u.firstname=:firstname, u.lastname=:lastname where u.id = :id")
+    void updateUser(@Param(value = "id") Integer id,   @Param(value = "firstname") String firstname, @Param(value = "lastname") String lastname );
 
 }
