@@ -22,13 +22,12 @@ public class UserService {
                 .username(userRequest.getUsername())
         .firstname(userRequest.getFirstname())
         .lastname(userRequest.lastname)
-        .country(userRequest.getCountry())
         .role(Role.USER)
         .build();
         if (!jwtService.isTokenValid(token,user)) {
             throw new Exception("No autorizado: el nombre de usuario no coincide.");
         }else {
-            userRepository.updateUser(user.id, user.firstname, user.lastname, user.country);
+            userRepository.updateUser(user.id, user.firstname, user.lastname);
             return new UserResponse("El usuario se registró satisfactoriamente");
     }}
 
@@ -42,7 +41,6 @@ public class UserService {
             .username(user.username)
             .firstname(user.firstname)
             .lastname(user.lastname)
-            .country(user.country)
             .build();
             return userDTO;
         }
