@@ -6,6 +6,7 @@ import { CabinListComponent } from './cabin-list/cabin-list.component';
 import { SearchCabinPageComponent } from './pages/search-cabin-page/search-cabin-page.component';
 import { CreateCabinComponent } from './create-cabin/create-cabin.component';
 import { CabinDetailComponent } from './cabin-detail/cabin-detail.component';
+import { CreateReservationComponent } from './create-reservation/create-reservation.component';
 
 export const routes: Routes = [
     {path:"login", component:LoginComponent},
@@ -14,5 +15,6 @@ export const routes: Routes = [
     {path:"home",component:HomeComponent},
     {path:"cabins",component:SearchCabinPageComponent},
     {path:"cabin/create",component:CreateCabinComponent},
-    {path:"cabin/:id",component:CabinDetailComponent}
+    {path:"cabin/:id",component:CabinDetailComponent},
+    {path:"reservation",component:CreateReservationComponent}
 ];
